command: Include deferral information in the plan result
If we've deferred any resources during the refresh and plan phases, these
will be tracked as part of the plan. To keep the user in the loop about
what's going on, we list out the deferred nodes and their associated
deferral reasons.

The user can either accept that Terraform will defer the indicated
resources, alter the configuration so that deferral is not necessary,
or use the -target option to customize what subset of resources are
applied such that the automatic deferral is not needed.